move datacloud mock data to mocks module
