+++ packages/template-retail-react-app/app/mocks/datacloud-mock-data.js
